+++ test/cli-unknown-argument.spec.ts
@@ -1,12 +1,10 @@
-import { match } from 'assert';
-import { bibtexTidy, test } from './utils';
+import { match, strictEqual } from 'assert';
+import { spawnSync } from 'child_process';
+import { BIN_PATH } from './targets/cli';
+import { test } from './utils';
 
 test('CLI should warn if an unknown argument is provided', async () => {
-	let err: unknown;
-	try {
-		await bibtexTidy([], {}, ['cli'], ['--foobar']);
-	} catch (e) {
-		err = e;
-	}
-	match(String(err), /Unknown option: --foobar/);
+	const proc = spawnSync(BIN_PATH, ['--foobar'], { encoding: 'utf8' });
+	strictEqual(proc.status, 1);
+	match(proc.stderr, /Unknown option: --foobar/);
 });